Очистил AppDelegate. Убрал из вью и добавил всё в отдельный класс
