+++ PhotoNotes/ViewController.swift
@@ -18,7 +18,6 @@ class ViewController: UIViewController{
     var imagePicker: ImagePicker!
     var defaultPicture = #imageLiteral(resourceName: "Image")
     private var cellViewModel = CellViewModel()
-    let locationManager = CLLocationManager()
     
     @IBOutlet weak var tableCellsView: UITableView!{
         didSet{
@@ -48,52 +47,9 @@ class ViewController: UIViewController{
     
     override func viewDidAppear(_ animated: Bool) {
         super.viewDidAppear(animated)
-        checkLocationEnabled()
+        //checkLocationEnabled()
     }
-    func checkLocationEnabled(){
-        if CLLocationManager.locationServicesEnabled() {
-            setupManager()
-        }
-        else{
-            
-            showAlertLocation(title: "Geolocation is unabled", message: "Do you want to turn it on?", url: URL(string: "App-prefs:root=LOCATION_SERVICES"))
-        }
-    }
-    func setupManager(){
-        locationManager.delegate = self
-        locationManager.desiredAccuracy = kCLLocationAccuracyBest
-    }
-    func checkAutorization(){
-        switch CLLocationManager.authorizationStatus() {
-        case .authorizedAlways:
-            break
-        case .authorizedWhenInUse:
-            locationManager.startUpdatingLocation()
-            break
-        case .denied:
-            showAlertLocation(title: "Geolocation is unabled", message: "Do you want to turn it on?", url: URL(string: UIApplication.openSettingsURLString))
-            break
-        case .restricted:
-            break
-        case .notDetermined:
-            locationManager.requestWhenInUseAuthorization()
-        }
-    }
-    func showAlertLocation(title:String,message:String?, url:URL? ){
-        let alert = UIAlertController(title: title, message: message, preferredStyle: .alert)
-        
-        let settingsAction = UIAlertAction(title: "Options", style: .default) { (alert) in
-            if let url = url{
-                UIApplication.shared.open(url, options: [:], completionHandler: nil)
-            }
-        }
-        let cancelAction = UIAlertAction(title: "Cancel", style: .cancel, handler: nil)
-        
-        alert.addAction(settingsAction)
-        alert.addAction(cancelAction)
-        
-        present(alert, animated: true, completion: nil)
-    }
+
     
 }
 
@@ -138,21 +94,9 @@ extension ViewController: ImagePickerDelegate{
         dateFormatter.timeStyle = .none
         dateFormatter.locale = Locale(identifier: "ru_RUS")
         dateFormatter.setLocalizedDateFormatFromTemplate("dd-YYYY-MM")
-        cellViewModel.createCell(imageNew: image ?? defaultPicture, textNew: dateFormatter.string(from: Date()),  realAddress:         locationManag.location?.description ?? "dad")
+        cellViewModel.createCell(imageNew: image ?? defaultPicture, textNew: dateFormatter.string(from: Date()),  realAddress: "dad")
     }
 }
 
-extension ViewController:CLLocationManagerDelegate{
-    
-    func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]) {
-        if let location = locations.last?.coordinate{
-            let region = MKCoordinateRegion(center: location, latitudinalMeters: 5000, longitudinalMeters: 5000)
-           // locationManager.region
-        }
-    }
-    func locationManager(_ manager: CLLocationManager, didChangeAuthorization status: CLAuthorizationStatus) {
-        checkAutorization()
-    }
-    
-}
+
